+++ crates/query-engine/metadata/src/metadata/database.rs
@@ -30,6 +30,16 @@ pub struct ComparisonOperators(pub BTreeMap<ScalarType, BTreeMap<String, Compari
 pub struct ComparisonOperator {
     pub operator_name: String,
     pub argument_type: ScalarType,
+
+    #[serde(default = "default_true")]
+    pub is_infix: bool,
+}
+
+/// This is quite unfortunate: https://github.com/serde-rs/serde/issues/368
+/// TL;DR: we can't set default literals for serde, so if we want 'is_infix' to
+/// default to 'true', we have to set its default as a function that returns 'true'.
+fn default_true() -> bool {
+    true
 }
 
 /// Mapping from a "table" name to its information.