+++ crates/query-engine/translation/src/translation/mutation/mod.rs
@@ -1,5 +1,4 @@
 pub mod delete;
 pub mod generate;
 pub mod translate;
-
 pub use translate::translate;